+++ semesterstat/crud/summary/__init__.py
@@ -1,0 +1,2 @@
+from .mainqueries import MainSummary
+from .subjectqueries import SubjectSummary
